addr fb: move selector that do not require an existing application to a separate file
